+++ Aquality.Selenium/Aquality.Selenium.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29009.5
+# Visual Studio Version 17
+VisualStudioVersion = 17.8.34309.116
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{43453468-EBA2-4800-B6CA-765FA722BA61}"
 EndProject
@@ -10,6 +10,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Aquality.Selenium.Tests", "tests\Aquality.Selenium.Tests\Aquality.Selenium.Tests.csproj", "{856AE8E1-31EF-49A1-B6EB-C72BF12432F8}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{999ED05C-B034-45A1-8F11-2B519C308250}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Aquality.Selenium.Support", "src\Aquality.Selenium.Support\Aquality.Selenium.Support.csproj", "{72EB3821-6AB9-465C-9B88-D878D677C754}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -25,6 +27,10 @@ Global
 		{856AE8E1-31EF-49A1-B6EB-C72BF12432F8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{856AE8E1-31EF-49A1-B6EB-C72BF12432F8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{856AE8E1-31EF-49A1-B6EB-C72BF12432F8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{72EB3821-6AB9-465C-9B88-D878D677C754}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{72EB3821-6AB9-465C-9B88-D878D677C754}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{72EB3821-6AB9-465C-9B88-D878D677C754}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{72EB3821-6AB9-465C-9B88-D878D677C754}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -32,6 +38,7 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{B0E3FDB8-CF47-40D2-AC0E-FB12F66D1F26} = {43453468-EBA2-4800-B6CA-765FA722BA61}
 		{856AE8E1-31EF-49A1-B6EB-C72BF12432F8} = {999ED05C-B034-45A1-8F11-2B519C308250}
+		{72EB3821-6AB9-465C-9B88-D878D677C754} = {43453468-EBA2-4800-B6CA-765FA722BA61}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {0AC8C847-FF57-4A72-8E22-F0947E17DFAF}
