+++ Aquality.Selenium/src/Aquality.Selenium.Support/Aquality.Selenium.Support.csproj
@@ -1,0 +1,50 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+	  <GeneratePackageOnBuild>false</GeneratePackageOnBuild>
+	  <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
+	  <Authors>aquality automation committers</Authors>
+	  <Company>aquality automation</Company>
+	  <Description>Support classes for Aquality.Selenium, such as ByImage locator</Description>
+	  <PackageDescription>Support classes for Aquality.Selenium, such as ByImage locator</PackageDescription>
+	  <PackageLicenseExpression></PackageLicenseExpression>
+	  <RepositoryUrl>https://github.com/aquality-automation/aquality-selenium-dotnet</RepositoryUrl>
+	  <RepositoryType>git</RepositoryType>
+	  <PackageTags>selenium webdriver browser automation image locator</PackageTags>
+	  <PackageLicenseFile>LICENSE</PackageLicenseFile>
+	  <PackageReadmeFile>README.md</PackageReadmeFile>
+	  <Copyright>Copyright 2024 Aquality Automation</Copyright>
+	  <IsPackable>true</IsPackable>
+  </PropertyGroup>
+
+  <PropertyGroup>
+	  <DocumentationFile>Aquality.Selenium.Support.xml</DocumentationFile>
+	  <NoWarn>1591</NoWarn>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Aquality.Selenium\Aquality.Selenium.csproj" />	
+  </ItemGroup>
+  <ItemGroup>
+	<None Include="..\..\..\LICENSE">
+	  <Pack>True</Pack>
+	  <PackagePath></PackagePath>
+	</None>
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="..\..\..\README.md">
+      <Pack>True</Pack>
+      <PackagePath></PackagePath>
+    </None>
+  </ItemGroup>
+  <ItemGroup>
+	<PackageReference Include="OpenCvSharp4.runtime.osx_arm64" Version="4.8.1-rc" />
+	<PackageReference Include="OpenCvSharp4" Version="4.10.0.20240616" />
+	<PackageReference Include="OpenCvSharp4.runtime.linux-arm" Version="4.10.0.20240616" />
+	<PackageReference Include="OpenCvSharp4.runtime.osx.10.15-x64" Version="4.6.0.20230105" />
+	<PackageReference Include="OpenCvSharp4.runtime.win" Version="4.10.0.20240616" />
+	<PackageReference Include="OpenCvSharp4_.runtime.ubuntu.20.04-x64" Version="4.10.0.20240616" />
+  </ItemGroup>	
+	
+</Project>
