+++ Aquality.Selenium/src/Aquality.Selenium.Support/Locators/ByImage.cs
@@ -8,7 +8,7 @@ using System.Collections.ObjectModel;
 using System.IO;
 using System.Linq;
 
-namespace Aquality.Selenium.Elements.Interfaces
+namespace Aquality.Selenium.Support.Locators
 {
     /// <summary>
     /// Locator to search elements by image.
