+++ Aquality.Selenium/tests/Aquality.Selenium.Tests/Integration/ImageLocatorTests.cs
@@ -1,5 +1,6 @@
 ﻿using Aquality.Selenium.Browsers;
 using Aquality.Selenium.Elements.Interfaces;
+using Aquality.Selenium.Support.Locators;
 using Aquality.Selenium.Tests.Integration.TestApp.TheInternet.Forms;
 using NUnit.Framework;
 using OpenQA.Selenium;
